backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
scikit-learn==1.3.2
PyPDF2==3.0.1
python-docx==1.1.0
werkzeug==3.0.1
numpy==1.26.2
